+++ src/main/java/team1/Daangn_Clonecoding/web/posting/PostingController.java
@@ -13,12 +13,11 @@ import team1.Daangn_Clonecoding.domain.posting.Posting;
 import team1.Daangn_Clonecoding.domain.posting.postingrepository.PostingRepository;
 import team1.Daangn_Clonecoding.domain.posting.postingservice.PostingService;
 import team1.Daangn_Clonecoding.web.SessionConst;
-import team1.Daangn_Clonecoding.domain.posting.dto.PostingDetailResponse;
+import team1.Daangn_Clonecoding.web.posting.dto.PostingDetailResponse;
 import team1.Daangn_Clonecoding.domain.posting.dto.PostingForm;
-import team1.Daangn_Clonecoding.domain.posting.dto.PostingResponse;
+import team1.Daangn_Clonecoding.web.posting.dto.PostingResponse;
 import team1.Daangn_Clonecoding.web.posting.dto.SimplePostingSuccessResponse;
 import team1.Daangn_Clonecoding.web.response.CommonResponse;
-import team1.Daangn_Clonecoding.web.response.Success;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -52,7 +51,9 @@ public class PostingController {
                                                           @SessionAttribute(SessionConst.LOGIN_MEMBER) Long memberId,
                                                       @PageableDefault(sort = "createdDate") Pageable pageable) {
 
-        return postingService.findPagingPosting(memberId, pageable);
+        Slice<Posting> paging = postingService.findPagingPosting(memberId, pageable);
+
+        return paging.map(PostingResponse::new); // Content PostingResponse 로 변화 후 반환
     }
 
     @GetMapping("/posting/{postingId}")
@@ -61,7 +62,9 @@ public class PostingController {
                                                        @SessionAttribute(SessionConst.LOGIN_MEMBER) Long memberId,
                                                    @PathVariable Long postingId) {
 
-        return postingService.findDetailPosting(postingId, memberId);
+        Posting posting = postingService.findDetailPosting(postingId);
+
+        return new PostingDetailResponse(posting, memberId); // Content PostingDetailResponse 로 변화 후 반환
     }
 
     @PatchMapping("/posting") //게시물 상태 구매완료로 변경
